+++ src/models/company/member/role/service/delete.rs
@@ -4,12 +4,12 @@ use crate::models::company::access::util::check_is_owner_with_err;
 use diesel::prelude::*;
 use uuid::Uuid;
 
-/// Delete company member role
+/// Удаляет роль участников компании.
 pub(crate) fn del_role_member(
     logged_user_uuid: &Uuid,
     data: &DelRoleMemberData,
     conn: &mut PgConnection
-) -> ServiceResult<i32> {
+) -> ServiceResult<usize> {
     use crate::schema::role_member_translate_list::dsl::*;
 
     check_is_owner_with_err(
@@ -27,15 +27,11 @@ pub(crate) fn del_role_member(
     match del_role {
         Ok(x) => {
             debug!("Completed delete role: {:#?}", x);
-
-            Ok(x as i32)
+            Ok(x)
         },
         Err(err) => {
             debug!("Error delete role: {:#?}", err);
-
-            Err(ServiceError::BadRequest(
-                "Error delete role".to_string()
-            ))
+            Err(ServiceError::BadRequest("Error delete role".to_string()))
         },
     }
 }
