+++ src/models/user/notification/service/delete.rs
@@ -2,12 +2,12 @@ use crate::errors::ServiceResult;
 use diesel::prelude::*;
 use uuid::Uuid;
 
-/// Delete notifications logged user
+/// Удаляет уведомление для пользователя.
 pub(crate) fn delete_notifications(
     logged_user_uuid: &Uuid,
     notifications_ids: &[i32],
     conn: &mut PgConnection,
-) -> ServiceResult<i32> {
+) -> ServiceResult<usize> {
     use crate::schema::notification_ref::dsl as notification_ref;
     use crate::schema::notification_to_user::dsl as notification_to_user;
 
@@ -27,5 +27,5 @@ pub(crate) fn delete_notifications(
 
     // debug!("fn delete_notification ={:?}", &delete_notification);
 
-    Ok(count_del as i32)
+    Ok(count_del)
 }
